+++ Repositories/ComputerRepository.cs
@@ -39,4 +39,19 @@ class ComputerRepository
 
         return computers;
     }
+
+    public void Save (Computer computer) //salva computer no bdd
+    {
+        var connection = new SqliteConnection (_databaseConfig.ConnectionString);
+        connection.Open();
+
+        var command = connection.CreateCommand();
+        command.CommandText = "INSERT INTO Computers VALUES($id, $ram, $processor)";
+        command.Parameters.AddWithValue("$id", computer.Id);
+        command.Parameters.AddWithValue("$ram", computer.Ram);
+        command.Parameters.AddWithValue("$processor", computer.Processor);
+
+        command.ExecuteNonQuery();
+        connection.Close();
+    }
 }
